Remove helper methods moving them to Event/Notification classes
