+++ lib/catarse_paypal_express.rb
@@ -2,6 +2,8 @@ require 'active_merchant'
 require "catarse_paypal_express/engine"
 require "catarse_paypal_express/gateway"
 require "catarse_paypal_express/contribution_actions"
+require "catarse_paypal_express/event"
+require "catarse_paypal_express/notification"
 require "catarse_paypal_express/payment"
 require "catarse_paypal_express/payment_setup"
 require "catarse_paypal_express/payment_checkout"
